requirements.txt
opencv-python
stag-python
numpy==1.26.4
TIME-python
matplotlib
ultralytics
